requirements.txt
flask==2.3.3
pytest==7.4.0
pytest-flask==1.2.0
flask-wtf==1.1.1
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==1.3.0
httpx==0.24.1
